[node-forge] Refactoring. Move types to separate files (the same as in origin repo)
