+++ types/node-forge/mgf1.d.ts
@@ -1,0 +1,5 @@
+declare module "node-forge" {
+    namespace mgf1 {
+        function create(any: any): any;
+    }
+}
