+++ types/node-forge/oids.d.ts
@@ -1,0 +1,10 @@
+declare module "node-forge" {
+    const oids: typeof pki.oids;
+
+    namespace pki {
+        interface oids {
+            [key: string]: string;
+        }
+        var oids: oids;
+    }
+}
