+++ types/node-forge/package.json
@@ -79,6 +79,10 @@
         {
             "name": "BendingBender      ",
             "githubUsername": "BendingBender"
+        },
+        {
+            "name": "Dmitry Semigradsky",
+            "githubUsername": "Semigradsky"
         }
     ]
 }
