+++ types/node-forge/pkcs1.d.ts
@@ -1,0 +1,4 @@
+declare module "node-forge" {
+    namespace pkcs1 {
+    }
+}
