+++ types/node-forge/pss.d.ts
@@ -1,0 +1,7 @@
+declare module "node-forge" {
+    namespace pss {
+        type PSS = any;
+
+        function create(any: any): PSS;
+    }
+}
