+++ types/node-forge/util.d.ts
@@ -1,0 +1,96 @@
+declare module "node-forge" {
+    namespace util {
+        function isArray(x: any): boolean;
+        function isArrayBuffer(x: any): boolean;
+        function isArrayBufferView(x: any): boolean;
+
+        interface ArrayBufferView {
+            buffer: ArrayBuffer;
+            byteLength: number;
+        }
+
+        type ByteBuffer = ByteStringBuffer;
+        class ByteStringBuffer {
+            constructor(bytes?: Bytes | ArrayBuffer | ArrayBufferView | ByteStringBuffer);
+            data: string;
+            read: number;
+            length(): number;
+            isEmpty(): boolean;
+            putByte(byte: Byte): ByteStringBuffer;
+            fillWithByte(byte: Byte, n: number): ByteStringBuffer;
+            putBytes(bytes: Bytes): ByteStringBuffer;
+            putString(str: string): ByteStringBuffer;
+            putInt16(int: number): ByteStringBuffer;
+            putInt24(int: number): ByteStringBuffer;
+            putInt32(int: number): ByteStringBuffer;
+            putInt16Le(int: number): ByteStringBuffer;
+            putInt24Le(int: number): ByteStringBuffer;
+            putInt32Le(int: number): ByteStringBuffer;
+            putInt(int: number, numOfBits: number): ByteStringBuffer;
+            putSignedInt(int: number, numOfBits: number): ByteStringBuffer;
+            putBuffer(buffer: ByteStringBuffer): ByteStringBuffer;
+            getByte(): number;
+            getInt16(): number;
+            getInt24(): number;
+            getInt32(): number;
+            getInt16Le(): number;
+            getInt24Le(): number;
+            getInt32Le(): number;
+            getInt(numOfBits: number): number;
+            getSignedInt(numOfBits: number): number;
+            getBytes(count?: number): Bytes;
+            bytes(count?: number): Bytes;
+            at(index: number): Byte;
+            setAt(index: number, byte: number): ByteStringBuffer;
+            last(): Byte;
+            copy(): ByteStringBuffer;
+            compact(): ByteStringBuffer;
+            clear(): ByteStringBuffer;
+            truncate(): ByteStringBuffer;
+            toHex(): Hex;
+            toString(): string;
+        }
+
+        function fillString(char: string, count: number): string;
+        function xorBytes(bytes1: string, bytes2: string, count: number): string;
+        function hexToBytes(hex: Hex): Bytes;
+        function bytesToHex(bytes: Bytes): Hex;
+        function int32ToBytes(int: number): Bytes;
+        function encode64(bytes: Bytes, maxline?: number): Base64;
+        function decode64(encoded: Base64): Bytes;
+        function encodeUtf8(str: string): Utf8;
+        function decodeUtf8(encoded: Utf8): string;
+
+        function createBuffer(): ByteBuffer;
+        function createBuffer(
+            input: Bytes | ArrayBuffer | ArrayBufferView | ByteStringBuffer,
+            encoding?: Encoding,
+        ): ByteBuffer;
+
+        namespace binary {
+            namespace raw {
+                function encode(x: Uint8Array): Bytes;
+                function decode(str: Bytes, output?: Uint8Array, offset?: number): Uint8Array;
+            }
+            namespace hex {
+                function encode(bytes: Bytes | ArrayBuffer | ArrayBufferView | ByteStringBuffer): Hex;
+                function decode(hex: Hex, output?: Uint8Array, offset?: number): Uint8Array;
+            }
+            namespace base64 {
+                function encode(input: Uint8Array, maxline?: number): Base64;
+                function decode(input: Base64, output?: Uint8Array, offset?: number): Uint8Array;
+            }
+        }
+
+        namespace text {
+            namespace utf8 {
+                function encode(str: string, output?: Uint8Array, offset?: number): Uint8Array;
+                function decode(bytes: Uint8Array): Utf8;
+            }
+            namespace utf16 {
+                function encode(str: string, output?: Uint8Array, offset?: number): Uint8Array;
+                function decode(bytes: Uint8Array): string;
+            }
+        }
+    }
+}
